feat(AttachmentEdit): add AttachmentEdit
I'm building this with the current version of the modal for the time being given that the rest of the assistant layout doesn't seem to be built yet. This means that the footer is slightly different, as are the close buttons, title, background color, etc. I am also using the current PF6 code editor, which looks different. If this winds up being the ultimate design, we can go back and customize this to match once decisions have been made.

Fixes https://github.com/patternfly/virtual-assistant/issues/57