feat(FileDropZone): Add drop zone for file upload purposes
Again, this deviates from the design significantly because the design isn't based on PatternFly 6. We are using PatternFly components where available for speedy development purposes. The only real big change here from the PatternFly demo is that the drop zone is hidden until engaged, and that we're not showing the status indicators. The design did not seem to include any handling here explicitly beyond what a successful upload looks like. We probably need a way to indicate a failure in a similar location.

Fixes https://github.com/patternfly/virtual-assistant/issues/59