fix(CodeModal): Restyle to match updated design
A decision was made to stick to the original design rather than PF6, so I'm attempting to match here. One exception is the copy to clipboard icon - PF does not easily let me swap this, and forking it is more involved/code-heavy. We can file an issue and address this as a follow-on. The dark modal background also seems a little too dark in context, and I'm having a tough time changing the margins on the Monaco editor. It's third-party with a lot of absolute positioning.