feat(MessageBar): Add attach menu variant
Added attach menu variation to menu bar and worked it into a new demo. I'm concerned that I had to make a lot of props optional with the way MessageBar is structured. Maybe it would be better for a more composable approach, where you put the children directly in the message bar as a user.