fix(AttachMenu): Try to fix latency issue and refactor
There's a bug in the PF6 code for application launcher, where if you open the menu and type too fast, search gives back the wrong results. I think this is due to the setTimeout function. I moved the ref into a useEffect so it only populates when the menu is fully open, allowing us to drop the async setTimeout function. Also adding a refactor for the filter function.
i