fix(MessageBox): Announce new messages
Messages should be announced to assistive devices. I added a prop to MessageBox to allow for this and updated our two demos with message sending so they make announcements. I also documented this in the Chatbot container docs. I checked this worked with VoiceOver, but I can't speak to any other tools.