fix(ChatbotHeaderSelectorDropdown/ChatbotHeaderOptionsDropdown): Should be keyboard navigable
The drawer's focus trap requires a listener for the escape key. Becuase it wraps the children, they also wind up triggering the listener. I am passing in a separate prop for closing so it doesn't matter if they trigger, and turning off the trigger also for the menu items. Escape is also used for leaving dropdowns, and we have those present in the panel if actions are enabled.