fix(TextMessage): Allow for more complex list output
In Composer AI, we were seeing bad list rendering when lists contained complex formatting. React-markdown seems to categorize the complex formatting as p tags. Since they all had an inline-block on them, this caused rendering issues when the p tag was in an li. This should fix it. I also brought in a silly, more complex example based on Composer AI.