feat(QuickStartsTile): Add quick starts tile for messages
Quick starts tile is based on the PF QuickStarts extension tile, with slightly reduced functionality (namely status tracking and shared context). The purpose of this card is simply to trigger a QuickStart in a different section of unrelated UI, so it accepts an onClick function and passes in its id when clicked.