chore(docs): Make Message expandable code links unique
Added rehype plugin so we could adjust the expandable code toggle links in our User and Bot Messages. Eric recommended that these are unique. Because the code blocks are in the same Message, we need to do this by modifying the syntax tree. I added this functionality and also added a warning in case people try to do conflicting behaviors.